修复不能使用mvn命令install的问题
