+++ dq-easy-cloud-parent/dq-easy-cloud-common/src/main/java/com/dq/easy/cloud/DqEasyCloudModelApplication.java
@@ -4,7 +4,6 @@ import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
 import org.springframework.boot.autoconfigure.orm.jpa.HibernateJpaAutoConfiguration;
-import org.springframework.boot.test.autoconfigure.orm.jpa.AutoConfigureDataJpa;
 import org.springframework.context.annotation.ComponentScan;
 
 import com.dq.easy.cloud.module.basic.constant.DqBaseComponentScan;
@@ -16,7 +15,6 @@ import com.dq.easy.cloud.module.basic.constant.DqBaseComponentScan;
  * @date 2018年3月18日 上午12:47:25
  */
 @SpringBootApplication(exclude={DataSourceAutoConfiguration.class,HibernateJpaAutoConfiguration.class})
-@AutoConfigureDataJpa
 @ComponentScan(basePackages = { DqBaseComponentScan.COM_DQ_EASY_CLOUD})
 public class DqEasyCloudModelApplication {
 	public static void main(String[] args) {
