+++ dq-easy-cloud-parent/dq-easy-cloud-common/src/main/java/com/dq/easy/cloud/module/common/generator/code/java/desc/DqJavaModifierDesc.java
@@ -1,6 +1,6 @@
 package com.dq.easy.cloud.module.common.generator.code.java.desc;
 
-import com.dq.easy.cloud.module.common.generator.code.base.constant.DqCodeGenerateConstant.DqIgnoreField.DqModifierMappingEnum;
+import com.dq.easy.cloud.module.common.generator.code.java.constant.DqCodeGenerateJavaConstant.DqIgnoreField.DqModifierMappingEnum;
 
 public class DqJavaModifierDesc {
 	private Integer code;
