+++ dq-easy-cloud-parent/dq-easy-cloud-common/src/main/java/com/dq/easy/cloud/module/common/generator/code/java/pojo/bo/DqGenerateJavaClassBO.java
@@ -3,9 +3,9 @@ package com.dq.easy.cloud.module.common.generator.code.java.pojo.bo;
 import java.util.ArrayList;
 import java.util.List;
 
-import com.dq.easy.cloud.module.common.generator.code.base.constant.DqCodeGenerateConstant.DqIgnoreField.DqModifierMappingEnum;
 import com.dq.easy.cloud.module.common.generator.code.base.pojo.desc.DqTemplateDesc;
 import com.dq.easy.cloud.module.common.generator.code.base.pojo.rule.DqGenerateRule;
+import com.dq.easy.cloud.module.common.generator.code.java.constant.DqCodeGenerateJavaConstant.DqIgnoreField.DqModifierMappingEnum;
 import com.dq.easy.cloud.module.common.generator.code.java.desc.DqJavaModifierDesc;
 import com.dq.easy.cloud.module.common.generator.code.java.pojo.dto.DqGenerateJavaBaseDTO;
