+++ dq-easy-cloud-parent/dq-easy-cloud-common/src/main/java/com/dq/easy/cloud/module/common/generator/code/xml/constant/DqCodeGenerateXmlConstant.java
@@ -1,0 +1,15 @@
+package com.dq.easy.cloud.module.common.generator.code.xml.constant;
+
+/**
+ * 
+ * <p>
+ * 代码生成器xml常量类
+ * </p>
+ *
+ *
+ * @author daiqi
+ * 创建时间    2018年3月28日 下午2:49:17
+ */
+public class DqCodeGenerateXmlConstant {
+	
+}
