+++ dq-easy-cloud-parent/dq-easy-cloud-common/src/main/java/com/dq/easy/cloud/module/config/amqp/DqAmqpConfig.java
@@ -5,7 +5,6 @@ import org.springframework.amqp.core.BindingBuilder;
 import org.springframework.amqp.core.FanoutExchange;
 import org.springframework.amqp.core.Queue;
 import org.springframework.amqp.core.TopicExchange;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Conditional;
 import org.springframework.context.annotation.Configuration;
@@ -18,7 +17,6 @@ import com.dq.easy.cloud.module.mq.amqp.utils.DqAmqpUtils;
 
 @Configuration
 @Conditional(DqAmqpConditional.class)
-@ConditionalOnBean
 public class DqAmqpConfig {
 	@Bean
     public Queue queueTest() {
