--- dq-easy-cloud-parent/dq-easy-cloud-common/src/main/java/com/dq/easy/user/controller/UserController.java
@@ -1,15 +1,0 @@
-package com.dq.easy.user.controller;
-
-import com.dq.easy.cloud.module.basic.controller.DqBaseController;
-import org.springframework.stereotype.Controller;
-
-/**
- * 描述：用户控制转发类控制转发类
- * @author THINK
- * @date 2018-03-28 10:07:37
- */
-@Controller(value = "userDAO")
-public class UserController extends DqBaseController{
-	
-	
-}
