--- dq-easy-cloud-parent/dq-easy-cloud-common/src/main/java/com/dq/easy/user/dao/UserDAO.java
@@ -1,11 +1,0 @@
-package com.dq.easy.user.dao;
-
-/**
- * 描述：用户数据处理接口
- * @author THINK
- * @date 2018-03-28 10:07:37
- */
-public interface UserDAO {
-	
-	
-}
