--- dq-easy-cloud-parent/dq-easy-cloud-common/src/main/java/com/dq/easy/user/dao/impl/UserDAOImpl.java
@@ -1,15 +1,0 @@
-package com.dq.easy.user.dao.impl;
-
-import com.dq.easy.cloud.module.basic.repository.DqBaseRepository;
-import org.springframework.stereotype.Repository;
-
-/**
- * 描述：用户数据处理实现类
- * @author THINK
- * @date 2018-03-28 10:07:37
- */
-@Repository(value = "userDAO")
-public class UserDAOImpl extends DqBaseRepository{
-	
-	
-}
