--- dq-easy-cloud-parent/dq-easy-cloud-common/src/main/java/com/dq/easy/user/pojo/dto/UserDTO.java
@@ -1,13 +1,0 @@
-package com.dq.easy.user.pojo.dto;
-
-import com.dq.easy.user.pojo.entity.UserEntity;
-
-/**
- * 描述：用户数据传输类
- * @author THINK
- * @date 2018-03-28 10:07:37
- */
-public class UserDTO extends UserEntity{
-	
-	
-}
