--- dq-easy-cloud-parent/dq-easy-cloud-common/src/main/java/com/dq/easy/user/pojo/query/UserQuery.java
@@ -1,13 +1,0 @@
-package com.dq.easy.user.pojo.query;
-
-import com.dq.easy.cloud.module.basic.pojo.query.DqBaseQuery;
-
-/**
- * 描述：用户查询类
- * @author THINK
- * @date 2018-03-28 10:07:37
- */
-public class UserQuery extends DqBaseQuery{
-	
-	
-}
