--- dq-easy-cloud-parent/dq-easy-cloud-common/src/main/java/com/dq/easy/user/srvice/UserService.java
@@ -1,11 +1,0 @@
-package com.dq.easy.user.srvice;
-
-/**
- * 描述：用户服务接口
- * @author THINK
- * @date 2018-03-28 10:07:37
- */
-public interface UserService {
-	
-	
-}
