+++ dq-easy-cloud-parent/dq-easy-cloud-common/src/main/resources/codetemplates/mybaties.ftl
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<RETVAL success='true'>
+	<Shop>
+		<sid>1</sid>
+		<name>北京鑫和易通贸易有限公司</name>
+	</Shop>
+	<Shop>
+		<sid>2</sid>
+		<name>张家口市金九福汽车服务有限公司</name>
+	</Shop>
+</RETVAL>
