+++ dq-easy-cloud-parent/dq-easy-cloud-consumer/src/main/java/com/dq/easy/cloud/DqEasyCloudComsumerApplication.java
@@ -2,13 +2,10 @@ package com.dq.easy.cloud;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.test.autoconfigure.orm.jpa.AutoConfigureDataJpa;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.context.annotation.PropertySource;
 
 
 @SpringBootApplication
-@AutoConfigureDataJpa
 //@PropertySource("classpath:application-consumer.properties")
 @EnableDiscoveryClient
 public class DqEasyCloudComsumerApplication {
