+++ dq-easy-cloud-parent/dq-easy-cloud-producer/src/main/java/com/dq/easy/cloud/DqEasyCloudProducerApplication.java
@@ -3,11 +3,9 @@ package com.dq.easy.cloud;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
 
 
 @SpringBootApplication
-@EnableJpaAuditing
 @EnableDiscoveryClient
 public class DqEasyCloudProducerApplication {
 	public static void main(String[] args) {
