+++ dq-easy-cloud-parent/dq-easy-cloud-producer/src/main/java/com/dq/easy/cloud/producer/user/UserRepository.java
@@ -1,0 +1,8 @@
+package com.dq.easy.cloud.producer.user;
+
+import com.dq.easy.cloud.module.basic.repository.DqBaseRepository;
+
+//@Repository
+public class UserRepository extends DqBaseRepository{
+	
+}
