+++ dq-easy-cloud-pay-parent/dq-easy-cloud-pay-zfb-service/src/main/java/com/dq/easy/cloud/pay/zfb/DqEasyCloudZfbPayServiceApplication.java
@@ -2,7 +2,6 @@ package com.dq.easy.cloud.pay.zfb;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.test.autoconfigure.orm.jpa.AutoConfigureDataJpa;
 import org.springframework.cloud.client.discovery.EnableDiscoveryClient;
 import org.springframework.context.annotation.ComponentScan;
 
@@ -10,7 +9,6 @@ import com.dq.easy.cloud.pay.common.base.constant.DqEasyCloudPayComponentScan;
 
 @SpringBootApplication
 @EnableDiscoveryClient
-@AutoConfigureDataJpa
 @ComponentScan(basePackages = { DqEasyCloudPayComponentScan.COM_DQ_EASY_CLOUD,
 		DqEasyCloudPayComponentScan.COM_DQ_EASY_CLOUD_PAY_ZFB })
 public class DqEasyCloudZfbPayServiceApplication {
